(cherry picked from commit 87118db45d0a2e4ad65c926edf21cd14626dc2d9)
* Conflicts:
*	ttnn/cpp/ttnn/operations/transformer/sdpa/device/kernels/compute/sdpa.cpp
