+++ ttnn/cpp/ttnn/operations/transformer/sdpa/device/kernels/dataflow/dataflow_common.hpp
@@ -1,0 +1,83 @@
+#include <cstdint>
+#include "dataflow_api.h"
+
+template <uint32_t tile_bytes, uint32_t num_readers>
+constexpr uint32_t get_barrier_read_threshold() {
+    return ((512 / num_readers) * (1024 + 128)) / tile_bytes;
+}
+
+template <uint32_t num_heads, uint32_t block_size_t, uint32_t Wt>
+uint32_t virtual_seq_tile_id_to_physical_tile_id(
+    uint32_t seq_tile_idx, uint32_t cur_head, const volatile tt_l1_ptr uint32_t* const page_table_ptr) {
+    // Given some index in the sequence tiles in range [0, max_seq_len_t]
+    // Return the physical tile id for that tile row
+    constexpr uint32_t block_stride = num_heads * block_size_t * Wt;
+    const uint32_t head_offset = cur_head * block_size_t * Wt;
+
+    const uint32_t virtual_block = seq_tile_idx / block_size_t;
+    const uint32_t physical_block = page_table_ptr[virtual_block];
+    const uint32_t block_row_offset = seq_tile_idx % block_size_t;
+    const uint32_t block_offset = block_row_offset * Wt;
+    return physical_block * block_stride + head_offset + block_offset;
+}
+
+class TensorTileShape {
+    uint32_t shape[4];
+    uint32_t strides[4];
+
+public:
+    // Constructor to initialize with 4D shape
+    TensorTileShape(uint32_t d0, uint32_t d1, uint32_t d2, uint32_t d3) {
+        shape[0] = d0;
+        shape[1] = d1;
+        shape[2] = d2;
+        shape[3] = d3;
+
+        // Calculate strides (row-major order)
+        strides[3] = 1;
+        strides[2] = strides[3] * shape[3];
+        strides[1] = strides[2] * shape[2];
+        strides[0] = strides[1] * shape[1];
+    }
+
+    // Get flattened index from 4D coordinates
+    uint32_t id_of(uint32_t i0, uint32_t i1, uint32_t i2, uint32_t i3) const {
+        return i0 * strides[0] + i1 * strides[1] + i2 * strides[2] + i3 * strides[3];
+    }
+};
+
+template <bool is_dram = true>
+void read_chunk(
+    const InterleavedAddrGenFast<is_dram>& reader,
+    const uint32_t cb_id,
+    uint32_t start_tile_id,
+    const uint32_t rows,
+    const uint32_t cols,
+    const uint32_t tile_bytes,
+    const uint32_t barrier_threshold,
+    const bool transpose = false) {
+    // Read Q chunk
+    const uint32_t num_tiles = rows * cols;
+    cb_reserve_back(cb_id, num_tiles);
+    const uint32_t base_write_ptr = get_write_ptr(cb_id);
+    uint32_t outer_ptr_stride = transpose ? tile_bytes : cols * tile_bytes;
+    uint32_t inner_ptr_stride = transpose ? tile_bytes * rows : tile_bytes;
+
+    uint32_t barrier_count = 0;
+    for (uint32_t row = 0; row < rows; ++row) {
+        uint32_t write_ptr = base_write_ptr + row * outer_ptr_stride;
+        for (uint32_t col = 0; col < cols; ++col) {
+            noc_async_read_tile(start_tile_id, reader, write_ptr);
+            start_tile_id += 1;
+            write_ptr += inner_ptr_stride;
+
+            if (++barrier_count == barrier_threshold) {
+                noc_async_read_barrier();
+                barrier_count = 0;
+            }
+        }
+    }
+    noc_async_read_barrier();
+
+    cb_push_back(cb_id, num_tiles);
+}
